refactor: adiciona index.ts nos módulos para centralizar importações públicas seguindo convenção de monolito modular
