+++ backend/src/modules/auth/index.ts
@@ -1,0 +1,2 @@
+export * as User_Repository from '../user/user.repository.js';
+export { userAuthSelect } from '../user/user.select.js';
