+++ backend/src/modules/auth/use-cases/passwordRecovery/verifyOtp/verifyOtp.service.ts
@@ -1,9 +1,9 @@
 import { transformForHash } from '../../../../../shared/utils/crypto/cryptoUtils.js';
 import { generateUUIDv4 } from '../../../../../shared/utils/uuid/uuidUtils.js';
 import { AppError } from '../../../../../shared/utils/error.js';
-import * as User_Repository from '../../../../user/user.repository.js';
 import * as Otp_Repository from '../../../repositories/codeOTP.repository.js';
 import * as Token_Repository from '../../../repositories/token.repository.js';
+import { User_Repository } from '../../../index.js';
 
 const verifyCodeService = async (code: string, email: string) => {
   const user = await User_Repository.findByEmail(email);
