+++ backend/src/modules/user/index.ts
@@ -1,0 +1,1 @@
+export * as VerifyAcc_Service from '../auth/use-cases/verifyAccount/verifyAcc.service.js';
